+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Designate Organization as donor. Refs UIORGS-383.
 * Settings for banking information. Refs UIORGS-391.
 * Implement organization's banking information form. Refs UIORGS-390.
+* Implement organization's banking information details view. Refs UIORGS-389.
 
 ## [5.0.0](https://github.com/folio-org/ui-organizations/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v4.0.0...v5.0.0)